proxy vars from agent config
